type of GfxElement->colortable was changed to UINT32 in MAME37b12, thus had to change pointercolortable[] to UINT32 too

+++ mess/vidhrdw/vic6560.c
@@ -108,7 +108,7 @@ static UINT16 backgroundcolor, framecolor, helpercolor, white, black;
 static UINT16 mono[2], monoinverted[2], multi[4], multiinverted[4];
 
 /* transparent, white, black */
-static unsigned short pointercolortable[3] =
+static UINT32 pointercolortable[3] =
 {0};
 static struct GfxLayout pointerlayout =
 {
